Add docker-compose to project

+++ sensors/models.py
@@ -3,4 +3,4 @@ from django.db import models
 class Sensor(models.Model):
     id = models.IntegerField(primary_key=True)
     station_id = models.IntegerField(blank=False)
-    value = models.IntegerField(blank=False)
+    value = models.FloatField(blank=False)
